+++ open-codegen/opengen/main.py
@@ -1,31 +1,33 @@
-import opengen as og
 import casadi.casadi as cs
+import opengen as og
 
-u = cs.SX.sym("u", 5)
-p = cs.SX.sym("p", 2)
-phi = og.functions.rosenbrock(u, p)
-c = cs.vertcat(1.5 * u[0] - u[1],
-               cs.fmax(0.0, u[2] - u[3] + 0.1))
-bounds = og.constraints.Ball2(None, 1.5)
-problem = og.builder.Problem(u, p, phi) \
-    .with_penalty_constraints(c)        \
-    .with_constraints(bounds)
-meta = og.config.OptimizerMeta()               \
-    .with_optimizer_name("potato")             \
-    .with_version('0.1.2')                     \
-    .with_licence('LGPLv3')
-ros_config = og.config.RosConfiguration()      \
-    .with_package_name("potato_optimizer")  \
-    .with_node_name("potato_controller")    \
-    .with_rate(35)                          \
-    .with_description("cool ROS node")
-local_open = '/home/chung/NetBeansProjects/RUST/optimization-engine/'
+u = cs.SX.sym("u", 5)                 # decision variable (nu = 5)
+p = cs.SX.sym("p", 2)                 # parameter (np = 2)
+phi = og.functions.rosenbrock(u, p)   # cost function
+
+segment_ids = [0, 1, 2, 3, 4]
+cstrs = [og.constraints.NoConstraints,
+         og.constraints.NoConstraints,
+         og.constraints.Ball2(None, 1.5),
+         og.constraints.NoConstraints,
+         og.constraints.NoConstraints]
+dim = 5
+bounds = og.constraints.CartesianProduct(segment_ids, cstrs)
+
+problem = og.builder.Problem(u, p, phi)  \
+        .with_constraints(bounds)
+
+meta = og.config.OptimizerMeta()                \
+    .with_version("0.0.0")                      \
+    .with_authors(["P. Sopasakis", "E. Fresk"]) \
+    .with_licence("CC4.0-By")                   \
+    .with_optimizer_name("potato")
+
 build_config = og.config.BuildConfiguration()  \
     .with_build_directory("my_optimizers")     \
     .with_build_mode(og.config.BuildConfiguration.RELEASE_MODE)  \
     .with_open_version('0.7.0-alpha.1') \
-    .with_tcp_interface_config() \
-    .with_ros(ros_config)
+    .with_tcp_interface_config()
 solver_config = og.config.SolverConfiguration()    \
     .with_tolerance(1e-5)                          \
     .with_delta_tolerance(1e-4)                    \
@@ -35,14 +37,15 @@ solver_config = og.config.SolverConfiguration()    \
 builder = og.builder.OpEnOptimizerBuilder(problem,
                                           meta,
                                           build_config,
-                                          solver_config).with_verbosity_level(3)
+                                          solver_config).with_verbosity_level(3).with_generate_not_build_flag(True)
+
 builder.build()
 
-
 o = og.tcp.OptimizerTcpManager('my_optimizers/potato')
-o.start()
-r = o.call([1.0, 50.0])
-if r.is_ok():
-    status = r.get()
-    print(status.solution)
-o.kill()
+# o.start()
+# r = o.call([1.0, 50.0])
+# if r.is_ok():
+#     status = r.get()
+#     print(status.solution)
+# o.kill()
+